+++ packages/core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.12.0](https://github.com/medly/medly-components/compare/@medly-components/core@6.11.2...@medly-components/core@6.12.0) (2022-06-02)
+
+
+### Features
+
+* **theme:** allow [@font-face](https://github.com/font-face) ascent and descent overrides ([#666](https://github.com/medly/medly-components/issues/666)) ([13c59bc](https://github.com/medly/medly-components/commit/13c59bc72ced964692779487e3998c756176480d))
+
+
+
+
+
 ## [6.11.2](https://github.com/medly/medly-components/compare/@medly-components/core@6.11.1...@medly-components/core@6.11.2) (2022-05-17)